+++ cc.config.json
@@ -54,7 +54,7 @@
             "overrides": {
                 "pal/minigame": "pal/minigame/non-minigame.ts",
                 "pal/audio": "pal/audio/web/player.ts",
-                "pal/sys": "pal/sys/web/index.ts"
+                "pal/system": "pal/system/web/system.ts"
             }
         },
         {
