+++ pal/minigame/xiaomi.ts
@@ -1,5 +1,5 @@
 import { IMiniGame } from 'pal/minigame';
-import { Orientation } from '../sys/enum-type/orientation';
+import { Orientation } from '../system/enum-type/orientation';
 import { cloneObject } from '../utils';
 
 declare let qg: any;
