+++ pal/system/enum-type/browser-type.ts
