+++ pal/system/enum-type/network-type.ts
