+++ pal/system/enum-type/operating-system.ts
