+++ pal/system/enum-type/platform.ts
