+++ tsconfig.json
@@ -32,7 +32,7 @@
         "./node_modules/@types/jest/index",
 
         // pal
-        "./@types/pal/sys",
+        "./@types/pal/system",
         "./@types/pal/minigame",
         "./@types/pal/audio"
       ]
